Package API level files (#576)
Summary:
Pull Request resolved: https://github.com/facebook/redex/pull/576

Generate a Python module for the API level files.

As the files together have nontrivial size (about 50MB at the time of writing), compress the files with zip or lzma, depending on availability in the host Python. This makes the output slightly nonportable in case lzma is supported on the generating host but not the target.

The generated file contains the compressed files as a base64 text blob and provides a minimal API to get the file content. For lzma compression, the file is about 900KB at time of writing - it is unclear how to increase the compression level with the Python API, as manual experiments show it should be compressible to about 700K. For zip compression, the file is around 8MB.

Let `redex.py` scan the input arguments. In case no API levels argument is given at all, attempt to load the generated module and extract its files. In case the module cannot be loaded, only log a warning message. This makes the approach more robust and corresponds to the original behavior.

Reviewed By: minjang

Differential Revision: D25314571

fbshipit-source-id: 0815809316410ed1464e8ecfb7a1ffc075e28ae8