Main: prologue + onMethodExit instrumentation only
Summary:
Finally, let's implement the main instrumentation. Three main parts:
1.  Allocation code: instrument prologue for register vectors and method offset.
2.  onMethodExit: instrument onMethodExit on possibly multiple exit blocks. Handle a large number of register vectors.
3. Block tracing: Instrument block coverage tracing computation code.

This diff implements the 1 and 2. In this diff, `MAX_BLOCKS` is zero, which is effectively identical to the method tracing.

Note that this diff is basically to break down the big implementation into smaller pieces for easier review and follow up. However, this diff still runs. FB4A is instrumented correctly and runs.

Reviewed By: agampe

Differential Revision: D25367042

fbshipit-source-id: cf711e817280acb5288ca231711ca06e22105cc5