More redundancy in DataFlowGraph representation.
Summary:
Previously, the edges only stored the details of the node they were not
accessed from (i.e. edges in inbound lists only stored sources, and edges in
outbound lists only stored targets).

Again, to simplify the implementation of the Sparta interface, which requires
implementations for sources and targets for any edge, the edge representation
has been grown so that regardless of the path they are accessed from, they can
identify their source and target node.

Reviewed By: newobj

Differential Revision: D25334653

fbshipit-source-id: 50b2de7be42f9cca4c2a35a3bc8bd2f3d446c2c3