Introduce DataFlowGraph
Summary:
`DataFlowGraph` is a mutable, symmetric representation of the kinds of
data-flow graphs that `MatchFlow` works with, where nodes are represented by
pairs of instructions and the abstract locations they are being interpreted as,
and edges are directed and labelled by the source operand that is being flowed
into.

`Locations` is a pre-existing representation of the same kind of graph, that is
suited to efficient queries, but is asymmetric (only keeps track of inbound
edges).  This is a problem when defining a graph interface for Sparta which
needs to be compute both predecessors (easy with `Locations`) and successors
difficult/inefficient with `Locations`.

Reviewed By: newobj

Differential Revision: D25334661

fbshipit-source-id: 0bc5005a3fd8c014bcb75316825bbe947711e54d