Universally quantified flow constraints
Summary:
Add a flow constraint flag to indicate that all instructions flowing into the
source must match its constraint (as opposed to the current `exists` default:
at least one instruction flowing into the source matches its constraint).

This has knock-on implications:

Because in well-formed IR, every operand has a source instruction, `forall`
implies `exists` (i.e. it cannot be vacuously true).

The interaction between `forall` and `alias`/`result` was not usable:
Previously, `alias`/`result` would apply constraints to moves and move-results
as it looked past them.  This makes it essentially impossible to use in
conjunction with `forall`, because the constraint would need to apply to both
the terminal (non-move) instruction as well as all the moves.

This diff also changes that behaviour:  Now it is not possible for an `alias`
or `result` modified constraint to match on the moves and move-results they
look through.

Reviewed By: minjang

Differential Revision: D25493757

fbshipit-source-id: 442fd4e9f5535f07f960d4a2df46e08bf2777d74