Introducing flag_t
Summary:
`flag_t` is used to modify what is meant by a flow constraint.  This diff
introduces the type, and gives names for the current defaults, and plumbs it
through.

The `flag_t` widget is a way to pass two flag parameters (a `AliasFlag` and a
`QuantFlag`), each with their own defaults, without those parameters being
sensitive to order, and without requiring the first parameter's default be
explicitly passed to modify the second.

Reviewed By: minjang

Differential Revision: D25462832

fbshipit-source-id: 8ce5a260daf9c5403466443dd22b3739d2631281