Reorder type lists to reduce number of protos (and shorty strings)
Summary:
This unifies superficially different protos, e.g.
```
    foo(int x, double y);
    bar(double x, int y);
```
would get transformed to
```
    foo(double y, int y);
    bar(double x, int y);
```

This safes space as it reduces the number of unique "protos" and "type lists" as well as "shorty" strings needed.

We only do the transformation for methods defined in the processed dexes. We do not apply the transformation to methods that have a proto that is defined externally or has a native implementation, or that cannot be renamed --- all those protos will stick around anyway, and there's no point in trying transform them where we can. (Actually, that's not quite true in the current implementation: We don't do the transformation if a proto reference will around in any dex; it might still be beneficial to do it in unaffected dexes.) The transformation decision is made at the proto-level, so all methods with the same proto get excluded from the transformation if any method is problematic.

Special care is applied to ensuring that we don't disturb virtual scopes, introducing a new renaming scheme based on the pair (name, original args).

Added new instrumentation tests.

Reviewed By: minjang

Differential Revision: D25338706

fbshipit-source-id: ffcbfa4cc773ae9311ca231e9ded6a983158b3f4