A test case for v-shaped predicates.
Summary:
The new test case showcases what happens when querying for results `matching` a
non-root location that appears in multiple flow constraints:  The instructions
matching the location are the ones that are reachable through any of those flow
constraints (rather than the stricter and more intuitive requirement that they
have to be reachable through all the flow constraints).

In the added test cases, there are four `const` instructions (call them `0` to
`3`).

 - `0` and `1` match the `lit` location because they are both flow into the
   respective operands of an `add-int` (which means that `add-int` matches
   `add`).
 - `2` also flows into an `add-int`, but because the other operand to the add
   is not another `const`, that `add-int` does not match `add`, so `2` is
   discarded from the results.
 - `3` matches the instruction constraint for `lit` but is otherwise not
   reachable, so is also not present in the results.

In the stricter interpretation, there would be no results, because no `const`
appears as both left and right operand to an `add-int`.

Reviewed By: minjang

Differential Revision: D25576452

fbshipit-source-id: 2eed3ad1658060c4e473ccbdef405280ba72a97d