unique should not increment past end.
Summary:
The `range` class' `unique` function unconditionally increments its `begin` to
check whether the result is equal to its `end`.  This is not always safe to do,
in particular for the empty range created by default constructing iterators,
this caused a seg fault.

Add a check that `begin` is not equal to `end` before incrementing it.

Reviewed By: Feng23

Differential Revision: D25587551

fbshipit-source-id: 0390d19ca6f7869406025c076b0c398837cd9229