Don't be afraid of invoke-super and invoke-interface in RemoveUninstantiablePass
Summary:
- There is nothing too special about invoke-interface, e.g. compared to the already handled invoke-virtual. We track "instantiable"ity of interface types just as all other types, otherwise other already established transformations in the same pass, e.g. the one for check-cast, wouldn't work.
- Even though I don't think that we should ever come across uninstantiable invoke-super instructions in `replace_uninstantiable_refs` (*), it would be okay to handle them like invoke-virtual, as no uninstantiable derived type can exist by construction of uninstantiable.

(*) At least as far as Java source code is concerned, any relevant invoke-super would have to come from a direct or virtual method in the same class or a derived class, which are all "uninstantiable", and thus its body gets reduced `replace_uninstantiable_refs` is never called. But this is a verification issue, and we are not going to fight it here, just optimize some more.

Reviewed By: minjang

Differential Revision: D25653082

fbshipit-source-id: aa39ad84adb1957952d26eb76850b001ef8e0cd4