Support python-internal xz
Summary:
Since we require Python 3.6, the `lzma` module is expected to be there. If the `xz` binary is missing, use the module for decompressing and compressing.

Retain the `xz` path as the primary method, as the Python module does not expose multi-threaded encoding and decoding, and thus may be slower.

Reviewed By: yuxuanchen1997

Differential Revision: D25717247

fbshipit-source-id: 00568e7d086c7368e3ee475c1b69aab9c0fea493