Fix missed block tracking if entry block is in a try-catch
Summary:
I checked the following post condition after block instrumentation:

```
num_instrumented_blocks == num_non_entry_blocks - num_rejected_blocks
```

However, I found a case where this post condition didn't hold. Interestingly, `num_instrumented_blocks` was always smaller by 1, which means we were missing one basic block for instrumentation.

After further debugging, all the cases were **when the entry block was in a try-catch.**

If so, when we instrument `onMethodBegin` in the entry block, due to the try-catch, it's creating an additional block for a possible throw. That's why we've missed one block for instrumentation.

The fix was quite simple. Don't remove entry block for instrumentation if the entry block is in a try-catch.

Reviewed By: agampe

Differential Revision: D25867027

fbshipit-source-id: 6680bd437978fa33431dda62c952b221cd96ba00