Only mark the ordered classes, but not the unordered classes of the primary dex as "perf sensitive".
Summary:
Together with the ISO change, this should be...
- a no-op for multidex APKs, and
- a small regression for APKs with `force_single_dex` where we now mark ordered cold-start classes as perf-sensitive.

Making the unordered classes in the primary dex to be not perf-sensitive allows for sorting them in the next diff up the stack. (Them being unordered means that we haven't observed them during startup in a meaningful way, so we don't know what order is best, so we might as well order them for best compression when the sort_remaining_classes option is provided. Again, all that one up the stack.)

Reviewed By: adicatana

Differential Revision: D25876477

fbshipit-source-id: b1e0545c31024d42cbc213c2b21d3a2c495d67e6