Register classes that are already in the starting dex with cross-dex-ref minimizer
Summary:
A few classes might have already been emitted to the "current" dex before cross-dex-ref minimization starts to fill up this dex with all the remaining classes, eventually spilling over into further dexes.

This diff makes it so that the minimizer knows that all the refs of those previously emitted classes in the "current" dex have already been emitted at the time cross-dex-ref minimization starts, so that it can start by choosing classes that have many of refs in common.

Reviewed By: minjang

Differential Revision: D25841900

fbshipit-source-id: 3f6e28e5cdbea4b16466c2a41fff3ef6d9e33fa4