Let cross-dex-ref minimization only pick "worst" class if we don't already have better refs
Summary:
At the "beginning" of each dex, the cross-dex-ref minimization find the "worst" remaining class --- that is, the class with the most refs. (The main purpose is to get some stability build-over-build.)

If the dex already has a bunch of classes in it that define more refs than the "worst" class, then it is beneficial to just continue the journey from there, picking the next best class in terms of refs, instead of adding another "worst" class. That's what this diff does.

Reviewed By: Feng23

Differential Revision: D25851144

fbshipit-source-id: b274e108e984598db2ad024bb4e70d86887394f9