final-"ish" fields in CSE
Summary:
We call a field final-"ish" when it is assigned once, and all uses happen after the assignment. In particular, the assignment doesn't need to happen in a constructor.

It's also okay for the object containing the field to be reused e.g. using the object-pool pattern, where then multiple single-assignment-followed-by-multiple-uses scenarios can occur, as long as no Redex analysis can possibly establish that it's the same re-used object.

In this case, CSE can treat the field just like regular final fields, allowing it to reuse previous reads for later ones. The actual mechanism is that no kind of memory barrier can invalidate a previous read final-"ish" field value.

There is currently no automatic inference for such fields, but this diffs makes it configurable via name matching.

Reviewed By: thezhangwei

Differential Revision: D25835654

fbshipit-source-id: c243381f63d8614e1f95ebd3cb98e2855b046939