Fix FrequentlyUsedPointers caching
Summary:
The cache was never populated as a copy of the initial empty cache was handed out by `RedexContext`.

This update puts the onus on the client to initialize the cache at the correct time. The advantage of said solution is that the cache does not have to become properly synchronized for multi-threaded use.

|**App size**|**Instructions(before)**|**Instructions(after)**|**Delta**|
|Small|3.414T|3.259T|-4.5%|
|Large|37.163T|36.000T|-3.2%|

Reviewed By: yuxuanchen1997

Differential Revision: D26027919

fbshipit-source-id: ec161bed4c6f3a5f6cd3f22756c96abee0f9447f