Make CtorFieldAnalyzer instance sensitive
Summary:
The purpose of `CtorFieldAnalyzer` is to infer the type mapping of instance fields at the end of the ctor. Based on that we can initialize the type domain of the ifields after the ctor without counting implicit nulls if unnecessary.

While doing that we made a mistake in `CtorFieldAnalyzer`, which is we do not track on which instance the field accesses touch. As a result, we might account field writes on instances other than `this`. The added test of this change illustrates this bug.

This change addresses this issue by adding a `ThisPointerEnvironment` to the `DexTypeEnvironment`. With that we know the registers holding `this` pointer and we only track field operations against `this` obj in `CtorFieldAnalyzer`.

Reviewed By: ssj933

Differential Revision: D26071504

fbshipit-source-id: 388c95ded3ca5e31baebc12077219d9a895fe8ac