Do not attempt to track "entrypoints" on-the-fly.
Summary:
`MatchFlow`'s find algorithm needs to know which nodes to start propagating
flow constraints from.  These are referred to as entrypoints and are identified
as the nodes without inbound edges.

Previously, these nodes were tracked as the graph was built, by looking for the
nodes that did not have further obligations on the source of their operands,
but this is not sufficient.  It misses nodes in the graph which end up without
inbound edges because all the instructions that could satisfy its obligations
do not because they do not meet the necessary instruction constraint.

Now we wait for the graph to be built before determining which nodes should be
treated as entrypoints.

Reviewed By: minjang

Differential Revision: D25851298

fbshipit-source-id: 0a84669e77bcefd2e3aba20843f6cdd30a1b1160