Minor change to compute_caller_nonrecursive_callees_by_stack_depth
Summary:
Instead of walking the callees in the (deterrministic) order that they were discovered during an earlier opcode walk, we now throw away unnecessary duplicates, and order by names for determinism in `compute_caller_nonrecursive_callees_by_stack_depth`.

This changes ever so slightly where we break (mutually) recursive call chains, but it has no effect on correctness.

The real purpose of this diff is make sure that the next diff in the stack does not change any metrics, establishing its correctness.

Reviewed By: agampe

Differential Revision: D26153897

fbshipit-source-id: 9d1c1e804e0c0ebd9765d9eb1dbe56f243674687