Different method for matching a lookup table to its enum.
Summary:
The previous method looked at the field's name which contained a mangled form
of the Enum's class name.  This method works for lookup tables for Java Enums,
(although it has to work around some interesting edge cases to do with inner
classes) but not for Kotlin Enums.

The new method (based on data-flow matching) works in both scenarios.

Reviewed By: Feng23

Differential Revision: D25644304

fbshipit-source-id: c8a49023722754364aeff2baee05fe1726acad52