option to emit all line info
Summary:
When investigating crashes, we have the runtime "line number" at which the crash occurred. The dexdump output contains mappings from bytecode offsets to these runtime "line numbers". The symbolicated version of the stack trace will turn a runtime "line number" into an actual line number. But an actual line number can correspond to multiple bytecode offsets. So it's not possible to go from an actual line number to the right bytecode offset. I'm not sure that the runtime "line number" is guaranteed to be unique, but it seems to be a lot better.

Before this change, I'd have to look at both symbolicated and unsymbolicated dexdump in order to understand where a crash occurred. With this change, I can symbolicate a dexdump while preserving all the line number information.

Reviewed By: agampe

Differential Revision: D26210650

fbshipit-source-id: fb4174ae2f20d1c0def653370ccefcc4a950ee99