Actually account for the plugin reserved referenced, even if not running
Summary:
As title.

NOTE: That we are still not running any plugins on non-root stores (TBD if we should do this for any of the plugins that actually touch them), but the main scope of this change is to account for references that are added by other passes that run after Interdex (not as a plugin).

Reviewed By: minjang

Differential Revision: D26240731

fbshipit-source-id: 53434380517f4329be2adabce010298baf40b3d9