Work around a register clobbering issue
Summary:
There is an interesting three-way issue currently:
* `InstructionLowering` converts a `check-cast`+`move-result-pseudo-object` to `move-object`+`check-cast`
* The register allocator ensures that the register is not live before the `check-cast` so that this works.
* `CopyPropagation` may use the clobbered register in a catch handler covering the `check-cast`.

For now, work around the issue by removing all aliasing state over a `THROW` edge when requested - which coincides with whether the register allocator has run. This has relatively low impact, as it is not a very common case.

Add a test.

Reviewed By: NTillmann

Differential Revision: D26237484

fbshipit-source-id: 0b4eb439610b6aa5fc4285310d4a06dd5f579495