Using a new and better way to check for try blocks in methods
Summary:
The old way we used to check if a method has `try` blocks was incorrect (an editable CFG removes `MFLOW_TRY` MethodItemEntries from the IR code).

Adding the correct way to test for that and modifying every place that needed to check that to use the new and correct method.

Reviewed By: agampe

Differential Revision: D26351612

fbshipit-source-id: 3fc1b68087e88eaa815bedc2ef5be3273ebe1faa