Make dispatch target scanning failure more lenient
Summary:
The pass targets a simple pattern where the overriding method of the interface method is always on the implementor class of the interface. We collects all the overriding methods to create an interface dispatch stub. If the input does not strictly conform this pattern, the pass will fail with an assertion error. This can be too restricted at times.

This change makes the handling of a non conforming pattern more lenient. If any of the dispatch target method is missing, we do not remove the interface and discard intermediate results for that interface type.

Reviewed By: yuxuanchen1997

Differential Revision: D26379168

fbshipit-source-id: 713061171e2cf469f8f4a331699d50f9053c40b0