Ensure that move-results are in the same chain
Summary:
The linearization of control flow graphs happens in ~three stages.
* First "chains" are created, which are runs of blocks connected by GOTOs. This keeps normal flow together, avoiding synthetic `goto` instructions, and also reduces the complexity of the second step.
* The second step computes a weak topological ordering over the chains and walks the ordering, emitting the blocks.
* Jumps, targets, trys and catches are inserted into the block order.

The WTO ordering does not enforce any semantical constraints on bytecode - including the single major one that `move-result` must immediately follow the "primary" instruction. This has to be handled at the chain stage.

However, the chain stage previously did *not* handle this correctly. It is possible to visit the block with `move-result` *before* the block with the primary instruction. An example of that is splitting the primary's block, which creates a *new* block with a higher number.

To handle this, when the `move-result` has already been handled, its chain needs to be merged into the primary's chain.

Reviewed By: amnn

Differential Revision: D26618718 (https://github.com/facebook/redex/commit/34a5fa7410e00670e92c472109fb59e6cd07ba85)

fbshipit-source-id: 4fe6abf3cf340bd351bd799e0ae7ff24574a0f80