Model native code in Redex
Summary: This diff starts to have Redex read the possible output of LLVM toolchain that's used to build native code. We are not using this information to perform any optimization yet. This diff is setting up the context that will support analysis and optimization enabled by having knowledge about the native code.

Reviewed By: lanza

Differential Revision: D26626348

fbshipit-source-id: 1079745b8a0c364f10c6a9d092044f4886cc55ad