cleanup old "enable_ab_experiments" configs
Summary:
Not a configuration anymore, the next diff shows how we'll setup experiments from now on. We will need to enable Rabbit only in stable and also pass, for the enabled experimental passes, what behaviour they should have.

If no config is passed for an experimental pass, it will just fallback on test (i.e., applying the changes).

Reviewed By: amnn

Differential Revision: D26667632

fbshipit-source-id: 5a64d2b999873d3059802d95473ac5dce1540e5f