Enable ABExperiments on Android 9 Arm64 Henosis builds
Summary:
ABExperimentation is turned off if its an art build.

For various reasons, Henosis builds are like art builds, so they've been getting that flag.
At the same time, Android 9/arm64 builds  are henosis builds, and we want to include them in experimentation.

Changing the logic for how we add the redacted flag to include henosis builds on android 9 and below, and use redacted flag to turn on ABExperimentation.

Reviewed By: agampe

Differential Revision: D26696144

fbshipit-source-id: ec6b796eeff48b74611d360fea5e1d2ee0d10206