Fix instrumentation ordering
Summary:
Glad that I've found this bug and fixed it.

The previous could resulted in like this. `>` denotes the code that we've inserted. Notice that `OR_INT` is executed after `onMethodExit`, which is wrong. Also notice that this OR operation is effectively dead.

```
Block B3:
  preds: (goto B2) (goto B7)
  POSITION: DynamicAnalysis;.verify:(I)V(DynamicAnalysis.java:139)
  SOURCE-BLOCK: DynamicAnalysis;.verify:(I)V@2
>  OPCODE: INVOKE_STATIC v7, v6, DynamicAnalysis;.onMethodExit:(IS)
>  OPCODE: OR_INT_LIT16 v6, v6, 2
  OPCODE: RETURN_VOID
```

The root cause was we pre-calculated the insertion point of `OR_INT`, before actual instrumentation. If a block has only `RETURN_VOID`, the insertion position will be the `RETURN_VOID`. But we add `INVOKE_STATIC` before the `OR_INT`. The late dead code removal removes this code, resulting in losing tracking this block.

This diff fixes the ordering. `v6` is updated being reported.
```
Block B3:
  preds: (goto B2) (goto B7)
  POSITION: DynamicAnalysis;.verify:(I)V(DynamicAnalysis.java:139)
  SOURCE-BLOCK: DynamicAnalysis;.verify:(I)V@2
  OPCODE: OR_INT_LIT16 v6, v6, 2
  OPCODE: INVOKE_STATIC v7, v6, DynamicAnalysis;.onMethodExit:(IS)
  OPCODE: RETURN_VOID
```

Note: there could be multiple exit blocks in a method. So, we still need to track which exit blocks were executed.

Reviewed By: agampe

Differential Revision: D26710126

fbshipit-source-id: 08f784338610e264ac8cca46413788ddce6ce576