Avoid GCC's `-Wformat-zero-length`
Summary:
We want `-Wformat` to avoid crashes in logging. But GCC enables a warning that is controversial and hits us hard: it warns about `""`, which we use for message-less assertions. Compiling right now gives 17745 warnings for this.

It is complicated to set up the warnings correctly so that neither GCC nor Clang complain (as the latter has no such warning, at all). So use `" "` instead as a sentinel.

Reviewed By: Feng23

Differential Revision: D26837447

fbshipit-source-id: 738d8e714c246fe7a796be1712ac639582e638df