Ignoring ACC_MODULE classes.
Summary:
Modules feature as a new kind of constant in the constant pool and a new access
modifier, as part of the Module/Package system in Java 9.  When Redex is fed a
JAR with this kind of class/constant, it will complain about two things:

 - It doesn't understand the constant pool tag.
 - The class with a `ACC_MODULE` access modifier doesn't have a superclass.

This diff adds support to recognise and then swiftly ignore the tag and
classes with that modifier.

Reviewed By: kuganv

Differential Revision: D26882779

fbshipit-source-id: 1e31c58cddc3884bcd305fe3f1bd7a70708e3dd4