Disable if Kotlin Intrinsics class is external.
Summary:
RemoveNullcheckStringArgPass adds extra methods to Kotlin's `Intrinsics` class,
wrapping methods that already exist there.  This does not work, however, if
the `Intrinsics` class is external (because then Redex is not free to modify
it).

Explicitly check for this case and abort the optimisation, to avoid introducing
calls to methods that we can't add.

Reviewed By: kuganv

Differential Revision: D26882780

fbshipit-source-id: 94a7926d85dc23a3d8c6d95eaf79677d3408637f