When analyzing impact of constant arguments, also simulate local-dce to further refine cost estimate of ultimately remaining code
Summary:
In the analysis to determine inlined code cost, we already used information from the constant-propagation analysis to ignore blocks that become unreachable for particular constant arguments. In this diff, we now also perform a (backwards) liveness analysis, exactly leveraging the local-dce code. Together, this allows us to precisely determine the pruning effects of constant propagation and local-dce, when they will later be applied on the inlined code during "shrinking" in the inliner.

This actually addresses a TODO that's been in the code for a while ("TODO: Simulate backwards LocalDCE run to find what instructions actually remain live").

Note that this change does not affect the correctness of the inlining, it just guides its heuristics.

Reviewed By: kuganv

Differential Revision: D26555009

fbshipit-source-id: 727850a22dc39b2f73d9b6ed243a238b7d77d195