Refining cross-dex penalty computation
Summary:
It used to be based on the full callee weight, not taking into account that we know more after applying a constant argument analysis.

We now use the average over what remains after constant argument analysis.

Note that this change only tweaks the inliner heuristics, and does not affect correctness.

On its own, this change doesn't do much, but it's important for the later diffs in the stack to really do their work.

Reviewed By: kuganv

Differential Revision: D26584048

fbshipit-source-id: 7a3e3bfa984d1b4d0f3d3ff8ad781901f86d0a22