encoding of outlined positions
Summary:
This adds a new `PositionPatternSwitchManager`, accessible and life-time managed via the RedexContext.
It is used by diffs up the stack.

This class manages...
- "position patterns". A position pattern represents a list of positions that may be applicable for an outlined method when invoked in a particular context.
- "position switches". A position switch represents a list of pattern-based position cases.

Special positions may refer to such patterns and switches. They are identified via certain magic class / method names:
- redex.$Position.pattern, with the pattern-id as its line number (going to be used at an outlined method callsite)
- redex.$Position.switch, with the switch-id (later rewritten to the line-offset of a counted case list) as its line number (going to be used at individual instructions within an outlined method)

And to represent counted case lists:
- redex.$Position.count, with the count of cases as its line number
- redex.$Position.case, with the pattern-id as its line number, and an actual position (chain) referenced by its parent field

For example, the invocation of an outlined method may be decorated with a particular pattern-id, e.g. 12345:
```
position: method Lredex/$Position;.pattern:()V, line 12345 (no parent)
instruction: invoke (outlined method)
```

Then, within the outlined method, positions can be represented by special switch positions such as the following, for some switch-id 99991:
```
method Lredex/$Position;.switch:()V, line 99991 (no parent)
```

We encode the switches and cases via extra positions in the v2-map file.  Later, when a v2-map file is created, the line number 99991 will be replaced with an offset to such a switch-case table, so the switch-ids used while Redex is running won't be found in the map file. In the following example, that offset would be 101, where we'll create 3 consecutive entries such as the following, where 12345 and 54321 are pattern-ids. What follows is a representation of the entries of the v2-map file.
```
...
23: (some actual position)
...
42: (some actual position)
...
101: method Lredex/$Position;.count:()V, line 2 (no parent)
102: method Lredex/$Position;.case:()V, line 12345, parent 23
103: method Lredex/$Position;.case:()V, line 54321, parent 42
```
The cases are sorted by their pattern-id, to enable doing a binary search to find a match.

Reviewed By: agampe

Differential Revision: D26949283

fbshipit-source-id: fc70e09687054a69fc4d9f0c48c1c16143c5d17b