Fix type tag field matching on multiple root types
Summary:
The way we tried to match type tag field with `MergerType` was not working properly.
We had a nested double loop that could cause mismatch between type tag field and the root of the `MergerType`.

This change fixes the bug by simply looping over all the `MergerType`s and finding the corresponding type tag field individually.

Reviewed By: NTillmann

Differential Revision: D27106386

fbshipit-source-id: 96b4fc9cde2779e1bce02680db9b888979cad1c7