Revert to synthetic position after pattern position when method has no debug positions
Summary:
That makes sure that the position of subsequent code isn't tainted by the outlined method pattern position.

I also sprinkled in a global boolean `CAN_OUTLINED_METHOD_INVOKE_OUTLINED_METHOD` as a guard to the limited functionality.
When disabled, as it is at this time, then we can now check via an assertion in `RealPositionMapper::process_pattern_switch_positions` that we truly never get nested outliner methods.

Reviewed By: ssj933

Differential Revision: D27175626

fbshipit-source-id: fa7c2ba8e6f929ba2b44db23570f7374f12eb7f0