consider classes for vestigial objects
Summary:
Recall that...
- we only remove writes to unread fields if we know that the stored values do not have a (relevant) object lifetime, or are "vestigial" objects, where
- we say an object is "vestigial" when the only references to it are stored in unread memory locations, not escaping in any other way, and if all objects it might hold in fields or array elements are vestigial objects.

(We only really care if an object can escape in a way where a weak reference to it is constructed, making its lifetime observable. As checking for that is difficult, we are conservative and consider any kind of escaping.)

In this diff, we generalize an earlier simple implementation that could only detect "vestigial" *arrays* to also handle "vestiginal" *objects* with fields.

In order to simplify testing, we also remove the otherwise not needed option to allow the "unsafe" removal of unread fields, which basically treated all objects as vestigial.

Reviewed By: amnn

Differential Revision: D27628320

fbshipit-source-id: fc3e4f6082ae8b2540281c4005e9f29fa520f096