Refactor BranchPrefixHoisting some more
Summary:
With the previous structure, skipping admissible blocks while looking for a second common function were not connected. This is actually important for the case where the next instruction is *not* a `move-result`, i.e., that there might be a `SourceBlock` in the middle.

Simplify `get_next_common_insn` to not be responsible for forwarding. Factor out the skipping of admissible checks into a lambda that also handles forward. Then let `get_insns_to_hoist` handle both.

Introduce a helper enum for the actual state of the iterators (as it is ternary). Mark with TODOs cases that could conceivably be optimized with some more checks (though a normal CFG should only exhibit one of the two).

Add more detailed tracing and a trace context so that the output for specific methods can be filtered easily.

Reviewed By: yuxuanchen1997

Differential Revision: D28368817

fbshipit-source-id: f2e120ff4300e3326973903525c9952c792a1314