Keep essential positions from empty blocks
Summary:
Retain positions when deleting an empty block and its successor block doesn't have any leading position of its own.

Background: Studying the "quality" of our debug positions, I noticed a problem where instructions following an outlined method invocation might be attributed to the special "pattern" position that's needed for symbolication of outlined code. This is wrong, as it might cause wrong attribution and/or triggers error in the symbolication pipeline. It may happen when the method we are outlining from didn't have any debug positions initially.
This is getting fixed here.

Reviewed By: beicy

Differential Revision: D28373141

fbshipit-source-id: 15068aca0544b34b4b48b2e017c223ea9752a374