Fix issue in block ordering
Summary:
When a CFG is constructed, the entry block will have the lowest ID, and will thus be handled first when building chains for linearization. If the entry block is changed, for example because it was empty, then other blocks will be handled first. It is possible that the entry block could then become a non-head part of a chain, if any block flows into it (e.g., if the entry is part of a loop). That would lead to linearization putting other blocks first.

Add a check that the result of `order()` always puts the entry block first. Update chain creation to always start with the entry block.

Add test cases.

Reviewed By: yuxuanchen1997

Differential Revision: D28429446

fbshipit-source-id: 68cb3f2a5fb7ee28d3dddd67760164548168ea17