Fix static build on macOS
Summary:
On macOS, we can never link to a static runtime since it is not provided, we have to link dynamically to the C standard library.
When building redex statically with `-BUILD_TYPE=Static`, our cmake script sets `Boost_USE_STATIC_RUNTIME=ON` and fails to find a boost variant.
This diff fixes the issue by keeping `Boost_USE_STATIC_RUNTIME` undefined.

Differential Revision: D28732611

fbshipit-source-id: 70e875b6640fff15df1fbc57e370a3b15f09781e