Fix static build on macOS
Summary:
On macOS, we can never link to a static runtime since it is not provided, we have to link dynamically to the C standard library.
When building redex statically with `-BUILD_TYPE=Static`, our cmake script sets `Boost_USE_STATIC_RUNTIME=ON` and fails to find a boost variant.
This diff fixes the issue by keeping `Boost_USE_STATIC_RUNTIME` undefined.

Reviewed By: dkgi

Differential Revision: D28732611

fbshipit-source-id: 43137544e9036e6dac14519d5846d839a95dc15b