Fix jsoncpp library path for static builds (#613)
Summary:
Pull Request resolved: https://github.com/facebook/redex/pull/613

When building redex with `-DBUILD_TYPE=Static`, we should be linking against static libraries (i.e, `.a` on macOS).
The problem is that the JsonCpp finder does not look for the static library and cmake silently goes on.
We end up getting a link time error with jsoncpp symbols missing.

This patch adds proper logic to find the static library for JsonCpp and error when it is not found.

Reviewed By: dkgi

Differential Revision: D28733019

fbshipit-source-id: 250cd4a28e148b9c887ba3e2b5251514f0f0dd3a