A few improvements
Summary:
to handle more cases.

VerticalMergingPass merges single-extended pairs. It merges parent class to its final single-extended child if only
the child type is used in type instructions. For example, there is only
`new Child` but no `new Parent`, child instances are never casted to parent instances,
then the parent class can be merged to child.

Previously the pass could not merge a parent class to a child when there are virtual
invocations with a ref from parent class like bellow.
```
invoke-virtual v0 Parent.method()
```
Since the instance must be the final child type but never be the parent type, we could
properly update the invocation.

Reviewed By: ssj933

Differential Revision: D28713827

fbshipit-source-id: 360656fab8746386b633fd0c72af0c83fdd5ef94