Rename ApkManager to AssetManager
Summary: This class seems to only exist to put files into the assets folder. Since we're going to work on .apk files and .aab files, the name `ApkManager` doesn't quite fit, so rename it (even though "AssetManager" is something specific within the Android SDK, I don't think it's that confusing to use it here too).

Reviewed By: ssj933

Differential Revision: D28917350

fbshipit-source-id: a4dda7b4364b948de5f749c881935aebe4d4c7ac