Matching single instructions in InstructionIterables and CFGs
Summary:
Two helper functions to extract single instructions that match a `match_t`
predicate, either from an iterable (of the kind returned by
`InstructionIterable` functions) or directly from a CFG.

These replace an existing helper of the same name, which worked over a vector
of `IRInstruction*`, which would typically need to be created via a
pre-processing step.

This diff also changes the interface to return the results in a vector rather
than accepting an out parameter -- can rely on RVO to clean this up for us.

Reviewed By: kuganv

Differential Revision: D28990047

fbshipit-source-id: bbba8b1aafa4f4d03ec1e62146ce33b58d5d51fe