intermediate remove-unreachable-blocks
Summary:
The inliner avoids creating methods that are too large by tracking `estimated_insn_size`.
While this should generally track closely the actual instructions size of the resulting method, there's some big discrepancy: when inlining a method with no return statement, it would not take into that large parts of the code might get cut off.

We earlier added an option to activate the "shrinker" to tighten the estimate even accounting for all subsequent local optimizations, but this turned out to be prohibitive in terms of Redex runtime.

This diff introduce a super cheap variation of this idea: We just run `remove_unreachable_blocks` on the cfg, which is basically a single pass over the cfg. This nicely captures the cut-off control-flow after inlining callees with no return (even when they were not going through the constant-propagation+local-dce simulation).

Reviewed By: beicy

Differential Revision: D29019207

fbshipit-source-id: ca5f1966bd77668ec814839e2e6dadc255d11473